+++ src/fields/SoSFBool.cpp
@@ -248,7 +248,7 @@ SoSFBool::convertTo(SoField * dest) const
   if (0) ;
 #if !defined(COIN_EXCLUDE_SOSFFLOAT)
   else if (dest->getTypeId()==SoSFFloat::getClassTypeId()) {
-    ((SoSFFloat *)dest)->setValue(this->getValue());
+    ((SoSFFloat *)dest)->setValue((float)this->getValue());
   }
 #endif // !COIN_EXCLUDE_SOSFFLOAT
 #if !defined(COIN_EXCLUDE_SOSFINT32)
